Add CookieView to every story
