+++ stories/1.1-standalone-custom.stories.tsx
@@ -6,21 +6,25 @@ import { docco } from 'react-syntax-highlighter/dist/esm/styles/hljs'
 
 // @ts-ignore
 import contents from 'raw-loader!./standalone-custom.html'
+import CookieView from './components/CookieView'
 
 const StandaloneConsentManagerExample = () => {
   return (
-    <Pane display="flex" height="calc(100vh - 10px)">
-      <iframe
-        frameBorder="0"
-        style={{ overflow: 'hidden', height: '100%', width: '550px', flex: 1 }}
-        height="100%"
-        width="550"
-        srcDoc={contents}
-      ></iframe>
-      <SyntaxHighlighter language="html" style={docco} customStyle={{ flex: 1 }}>
-        {contents}
-      </SyntaxHighlighter>
-    </Pane>
+    <>
+      <Pane display="flex" height="calc(100vh - 10px)">
+        <iframe
+          frameBorder="0"
+          style={{ overflow: 'hidden', height: '100%', width: '550px', flex: 1 }}
+          height="100%"
+          width="550"
+          srcDoc={contents}
+        ></iframe>
+        <SyntaxHighlighter language="html" style={docco} customStyle={{ flex: 1 }}>
+          {contents}
+        </SyntaxHighlighter>
+      </Pane>
+      <CookieView />
+    </>
   )
 }
